+++ lib/Service/RecipeService.php
@@ -919,13 +919,30 @@ class RecipeService {
 		return $this->db->findAllCategories($this->user_id);
 	}
 
+
+
+	/** Adds modification and creation date to each recipe in the list
+	 *
+	 * @param array $recipes
+	 */
+	private function addDatesToRecipes(array &$recipes) {
+		foreach ($recipes as $i => $recipe) {
+			// TODO Add data to database instead of reading from files
+			$r = $this->getRecipeById($recipe['recipe_id']);
+			$recipes[$i]['dateCreated'] = $r['dateCreated'];
+			$recipes[$i]['dateModified'] = $r['dateModified'];
+		}
+	}
+
 	/**
 	 * Gets all recipes from the index
 	 *
 	 * @return array
 	 */
-	public function getAllRecipesInSearchIndex() {
-		return $this->db->findAllRecipes($this->user_id);
+	public function getAllRecipesInSearchIndex(): array {
+		$recipes = $this->db->findAllRecipes($this->user_id);
+		$this->addDatesToRecipes($recipes);
+		return $recipes;
 	}
 
 	/**
@@ -935,8 +952,10 @@ class RecipeService {
 	 *
 	 * @return array
 	 */
-	public function getRecipesByCategory($category) {
-		return $this->db->getRecipesByCategory($category, $this->user_id);
+	public function getRecipesByCategory($category): array {
+		$recipes = $this->db->getRecipesByCategory($category, $this->user_id);
+		$this->addDatesToRecipes($recipes);
+		return $recipes;
 	}
 
 	/**
@@ -945,19 +964,22 @@ class RecipeService {
 	 * @param string $keywords Keywords/tags as a comma-separated string.
 	 *
 	 * @return array
+	 * @throws \OCP\AppFramework\Db\DoesNotExistException
 	 */
-	public function getRecipesByKeywords($keywords) {
-		return $this->db->getRecipesByKeywords($keywords, $this->user_id);
+	public function getRecipesByKeywords($keywords): array {
+		$recipes = $this->db->getRecipesByKeywords($keywords, $this->user_id);
+		$this->addDatesToRecipes($recipes);
+		return $recipes;
 	}
 
 	/**
 	 * Search for recipes by keywords
 	 *
-	 * @param string $keywords
-	 *
+	 * @param $keywords_string
 	 * @return array
+	 * @throws \OCP\AppFramework\Db\DoesNotExistException
 	 */
-	public function findRecipesInSearchIndex($keywords_string) {
+	public function findRecipesInSearchIndex($keywords_string): array {
 		$keywords_string = strtolower($keywords_string);
 		$keywords_array = [];
 		preg_match_all('/[^ ,]+/', $keywords_string, $keywords_array);
@@ -966,7 +988,9 @@ class RecipeService {
 			$keywords_array = $keywords_array[0];
 		}
 
-		return $this->db->findRecipes($keywords_array, $this->user_id);
+		$recipes = $this->db->findRecipes($keywords_array, $this->user_id);
+		$this->addDatesToRecipes($recipes);
+		return $recipes;
 	}
 
 	/**